refactor: migrate CategoriesPage to React Query with useFacetsQuery hook
