+++ src/pages/CategoriesPage.tsx
@@ -1,56 +1,8 @@
-import { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
-import { artRegistry } from "../services/registry";
-import type { FacetItem } from "../services/types";
-
-interface FacetSection {
-  label: string;
-  facet: "culture" | "classification" | "century";
-  items: FacetItem[];
-}
+import { useFacetsQuery } from "../hooks/useFacetsQuery";
 
 export default function CategoriesPage() {
-  const [sections, setSections] = useState<FacetSection[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-
-  useEffect(() => {
-    let cancelled = false;
-
-    async function loadFacets() {
-      setIsLoading(true);
-      setError(null);
-
-      try {
-        const [cultures, classifications, centuries] = await Promise.all([
-          artRegistry.fetchFacet("culture", 40),
-          artRegistry.fetchFacet("classification", 40),
-          artRegistry.fetchFacet("century", 30),
-        ]);
-
-        if (cancelled) return;
-
-        setSections([
-          { label: "By Culture", facet: "culture", items: cultures },
-          { label: "By Classification", facet: "classification", items: classifications },
-          { label: "By Century", facet: "century", items: centuries },
-        ]);
-      } catch {
-        if (!cancelled) {
-          setError("Failed to load categories. Please try again.");
-        }
-      } finally {
-        if (!cancelled) {
-          setIsLoading(false);
-        }
-      }
-    }
-
-    loadFacets();
-    return () => {
-      cancelled = true;
-    };
-  }, []);
+  const { data: sections, isLoading, error } = useFacetsQuery();
 
   return (
     <div className="categories-page">
@@ -66,10 +18,10 @@ export default function CategoriesPage() {
       )}
 
       {error && (
-        <div className="categories-page__status">{error}</div>
+        <div className="categories-page__status">Failed to load categories. Please try again.</div>
       )}
 
-      {!isLoading && !error && sections.map((section) => (
+      {sections?.map((section) => (
         <section key={section.facet} className="categories-page__section">
           <h2 className="categories-page__section-title">{section.label}</h2>
           <div className="categories-page__chips">
